+++ src/homer_lib/hmr_private.h
@@ -1036,12 +1036,12 @@ typedef struct hvenc_enc_t hvenc_enc_t;
 typedef struct low_level_funcs_t low_level_funcs_t;
 struct low_level_funcs_t
 {
-	uint32_t (*sad)(uint8_t * src, uint32_t src_stride, int16_t * pred, uint32_t pred_stride, int size);
-	uint32_t (*ssd)(uint8_t * src, uint32_t src_stride, int16_t * pred, uint32_t pred_stride, int size);
+	uint32_t (*sad)(int16_t * src, uint32_t src_stride, int16_t * pred, uint32_t pred_stride, int size);
+	uint32_t (*ssd)(int16_t * src, uint32_t src_stride, int16_t * pred, uint32_t pred_stride, int size);
 	uint32_t (*ssd16b)(int16_t * src, uint32_t src_stride, int16_t * pred, uint32_t pred_stride, int size);
-	void (*predict)(uint8_t * orig, int orig_stride, int16_t* pred, int pred_stride, int16_t * residual, int residual_stride, int size);
+	void (*predict)(int16_t * orig, int orig_stride, int16_t* pred, int pred_stride, int16_t * residual, int residual_stride, int size);
 	void (*reconst)(int16_t* pred, int pred_stride, int16_t * residual, int residual_stride, int16_t* decoded, int decoded_stride, int size);
-	uint32_t (*modified_variance)(uint8_t *p, int size, int stride, int modif);
+	uint32_t (*modified_variance)(int16_t *p, int size, int stride, int modif);
 
 	void (*create_intra_planar_prediction)(henc_thread_t* et, int16_t *prediction, int pred_stride, int16_t  *adi_pred_buff, int adi_size, int cu_size, int cu_size_shift);
 	void (*create_intra_angular_prediction)(henc_thread_t* et, ctu_info_t* ctu, int16_t *prediction, int pred_stride, int16_t  *adi_pred_buff, int adi_size, int cu_size, int cu_mode, int is_luma);
